requirements.txt
swibots
python-decouple
pypdf2
pillow